Renumber supp_figure_x_alsf_manual_annotations to supp_figure_4_alsf_manual_annotations
