+++ docs/manuscripts/cosmicqc_paper_v1/supp_figure_4_alsf_manual_annotations/figures/annotator_comparison_confusion_matrix_facet_celltype.png
